app/requirements.txt
requests
urllib3==1.26.6
pymongo==3.11.4
uvicorn==0.14.0
fastapi==0.66.0
pydantic==1.8.2
motor==2.4.0
config
pdfplumber==0.5.28
pandas
python-dotenv
pytz
git+https://github.com/mirea-ninja/rtu-schedule-parser@main